+++ GoWHunter.Fiddler/GoWHunter.Fiddler.csproj
@@ -62,6 +62,9 @@
   <ItemGroup>
     <Compile Include="LiteFiddlerApp.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="ReplaceConfig.cs" />
+    <Compile Include="ReplaceItem.cs" />
+    <Compile Include="ReplaceSet.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="packages.config" />
