+++ GoWHunter.Fiddler/ReplaceItem.cs
@@ -1,0 +1,9 @@
+namespace GoWHunter
+{
+    public class ReplaceItem
+    {
+        public string RegEx { get; set; }
+        public string Keyword { get; set; }
+        public string ReplaceWith { get; set; }
+    }
+}
