+++ GoWHunter.Fiddler/ReplaceSet.cs
@@ -1,0 +1,10 @@
+﻿using System.Collections.Generic;
+
+namespace GoWHunter
+{
+    public class ReplaceSet
+    {
+        public string Url { get; set; }
+        public List<ReplaceItem> ReplaceItems { get; set; }
+    }
+}
